+++ scripts/pod_lib_lint.rb
@@ -38,6 +38,8 @@ def main(args)
     exit(1)
   end
 
+  STDOUT.sync = true
+
   command = %w(bundle exec pod lib lint --sources=https://cdn.cocoapods.org/)
 
   # Figure out which dependencies are local
@@ -48,7 +50,21 @@ def main(args)
 
   command.push(*args)
   puts command.join(' ')
-  exec(*command)
+
+  # Run the lib lint command in a thread.
+  t = Thread.new do
+    system(*command)
+  end
+
+  # Print every minute since linting can run for >10m without output.
+  number_of_times_checked = 0
+  while t.alive? do
+    sleep 1.0
+    number_of_times_checked += 1
+    if (number_of_times_checked % 60) == 0 then
+      puts "Still working, running for #{number_of_times_checked / 60}min."
+    end
+  end
 end
 
 # Loads all the specs (inclusing subspecs) from the given podspec file.